Extract MCP tool dispatcher
